fix casched repository now works without connection to redis db

+++ src/main/java/indie/outsource/repositories/redis/CachedMongoDBProductRepository.java
@@ -16,53 +16,75 @@ public class CachedMongoDBProductRepository implements ProductRepository {
 
     private final MongoDatabase db = new DefaultMongoConnection().getDatabase();
     private final ProductMongoDbRepository mongoRepository = new ProductMongoDbRepository(db);
-    private final ProductRedisRepository redisRepository = new ProductRedisRepository();
+    private ProductRedisRepository redisRepository;
+
+    public CachedMongoDBProductRepository(){
+        try{
+            redisRepository = new ProductRedisRepository();
+        }
+        catch (Exception e){
+            redisRepository = null;
+        }
+    }
 
     @Override
     public void decreaseProductQuantity(ProductWithInfo product, int quantity) {
         mongoRepository.decreaseProductQuantity(product, quantity);
         ProductWithInfoDoc productDoc = ProductWithInfoMapper.fromDomainModel(mongoRepository.getById(product.getId()));
-        redisRepository.add(productDoc);
+        if(redisRepository != null){
+            redisRepository.add(productDoc);
+        }
     }
 
     @Override
     public void increaseProductQuantity(ProductWithInfo product, int quantity) {
         mongoRepository.increaseProductQuantity(product, quantity);
         ProductWithInfoDoc productDoc = ProductWithInfoMapper.fromDomainModel(mongoRepository.getById(product.getId()));
-        redisRepository.add(productDoc);
+        if(redisRepository != null){
+            redisRepository.add(productDoc);
+        }
     }
 
     @Override
     public List<ProductWithInfo> findAll() {
-        List<ProductWithInfoDoc> redisProducts = redisRepository.findAll();
-        if(redisProducts.isEmpty()) {
-            return mongoRepository.findAll();
+        if(redisRepository != null){
+            List<ProductWithInfoDoc> redisProducts = redisRepository.findAll();
+            if(!redisProducts.isEmpty()) {
+                return redisProducts.stream().map(ProductWithInfoDoc::toDomainModel).toList();
+            }
         }
-        return redisProducts.stream().map(ProductWithInfoDoc::toDomainModel).toList();
+        return mongoRepository.findAll();
     }
 
     @Override
     public ProductWithInfo getById(UUID id) {
-        ProductWithInfoDoc redisProduct = redisRepository.findById(id);
-        if (redisProduct == null) {
+        if(redisRepository != null) {
+            ProductWithInfoDoc redisProduct = redisRepository.findById(id);
+            if (redisProduct != null) {
+                return redisProduct.toDomainModel();
+            }
             ProductWithInfo mongoProduct = mongoRepository.getById(id);
             redisRepository.add(ProductWithInfoMapper.fromDomainModel(mongoProduct));
             return mongoProduct;
         }
-        return redisProduct.toDomainModel();
+        return mongoRepository.getById(id);
     }
 
     @Override
     public ProductWithInfo add(ProductWithInfo productWithInfo) {
         mongoRepository.add(productWithInfo);
         ProductWithInfo product = mongoRepository.getById(productWithInfo.getId());
-        redisRepository.add(ProductWithInfoMapper.fromDomainModel(product));
+        if(redisRepository != null){
+            redisRepository.add(ProductWithInfoMapper.fromDomainModel(product));
+        }
         return product;
     }
 
     @Override
     public void remove(ProductWithInfo productWithInfo) {
-        redisRepository.remove(ProductWithInfoMapper.fromDomainModel(productWithInfo));
+        if(redisRepository != null){
+            redisRepository.remove(ProductWithInfoMapper.fromDomainModel(productWithInfo));
+        }
         mongoRepository.remove(productWithInfo);
     }
 }
